test: commonize DB API integration fixtures and row counter
